Create linregEbModelSelVsN.py
This script is to recreate the figures produced by linregEbModelSelVsN.m. That matlab file relies on the pmtk3 toolbox, some of which has been recreated here. Those functions, included in this script, are:

1. linregFitEB - function to fit a linear regression with EB.
2. preprocessorApplyToTrain - function for applying preprocessing to data
3. linregPredictBayes - function to produce the posterior predictive given the output from linregFitBayes 
4. linregFitBayes - function to deploy which fitted functions should be applied to the data given the options.

In the toolbox, these functions are built out with more options than reflected in the script - most of those options weren't necessary for reproducing the graphs. I have written them such that they can be extended to the form of the functions seen in the toolbox.

Ultimately, I don't expect this script is the intended place for those functions. Since there isn't an analogous toolbox built out for the python side, I'm guessing the sanest place to put them is closest to where they are used.