Create linregRbfDemo.py
This is a python rewrite of linregRbfDemo.m. The first row of the graphic is different from that which appears in the textbook. In the textbook, the first row shows a perfect interpolation of 21 points using a design matrix of rank 10, which I believe is a mistake. With this script, we forego the perfect interpolation and keep in line with the description in the book. To make the figures look similar to the textbook graph otherwise, I changed the RBF width (so the caption in the book should change). Also, for the same purpose, some of the axis ranges were changed.