Create dirichlet3dPlot.py
As this script is right now, it produce the same angle seen in the book. Because of how python renders the image, one of the peaks looks more blue than it should (and I don't know how to resolve it :/ ). So I would suggest uncommenting line 36 for what I believe is a better angle.