Create mnist1NNdemo.py
On my computer, this has a 3.09% error rate in 185 seconds. Not bad!

This script anticipates that they'll be a MNIST folder within the data folder that contains the 4 MNIST data files. I can't (or just can't figure out how to) create that folder, so I have just uploaded those 4 files into the data folder.